Force cast double to int to correct ClassCastException

+++ ugs-platform/ugs-platform-surfacescanner/src/main/java/com/willwinder/ugs/platform/surfacescanner/AutoLevelerTopComponent.java
@@ -139,7 +139,7 @@ public final class AutoLevelerTopComponent extends TopComponent implements ItemL
 
         AutoLevelSettings als = settings.getAutoLevelSettings();
         this.stepResolution.setValue(als.stepResolution);
-        this.zSurface.setValue(als.zSurface);
+        this.zSurface.setValue((int) als.zSurface);
 
         this.bulkChanges = false;
         this.stateChanged(null);
